move RowParser to schemas_context
